feat(ui): hide user menu on docker extension [EE-4115]
closes [EE-4115]
